refactor(command)!: restructure classes placement
