+++ yui-adapter-command/src/main/java/me/whereareiam/yui/adapter/command/DefaultCommandService.java
@@ -4,9 +4,9 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import me.whereareiam.yui.adapter.command.definition.CommandDefinitionRegistry;
 import me.whereareiam.yui.adapter.command.definition.DefinitionProviderRegistry;
-import me.whereareiam.yui.adapter.command.registration.CommandDefinitionParser;
+import me.whereareiam.yui.adapter.command.parsing.definition.CommandDefinitionParser;
 import me.whereareiam.yui.adapter.command.registration.CommandScanner;
-import me.whereareiam.yui.adapter.command.registration.annotation.AnnotationCommandRegistrar;
+import me.whereareiam.yui.adapter.command.registration.AnnotationCommandRegistrar;
 import me.whereareiam.yui.model.command.CommandDefinition;
 import me.whereareiam.yui.service.CommandService;
 import me.whereareiam.yui.DefinitionProvider;
