+++ yui-adapter-command/src/test/java/me/whereareiam/yui/adapter/command/CommandRegistrationLifecycleTest.java
@@ -3,8 +3,8 @@ package me.whereareiam.yui.adapter.command;
 import me.whereareiam.yui.DefinitionProvider;
 import me.whereareiam.yui.adapter.command.definition.CommandDefinitionRegistry;
 import me.whereareiam.yui.adapter.command.definition.DefinitionProviderRegistry;
-import me.whereareiam.yui.adapter.command.registration.CommandDefinitionParser;
-import me.whereareiam.yui.adapter.command.registration.annotation.AnnotationCommandRegistrar;
+import me.whereareiam.yui.adapter.command.parsing.definition.CommandDefinitionParser;
+import me.whereareiam.yui.adapter.command.registration.AnnotationCommandRegistrar;
 import me.whereareiam.yui.annotation.command.Command;
 import me.whereareiam.yui.annotation.command.Definition;
 import me.whereareiam.yui.model.command.CommandDefinition;
